Initial `CurrentContent` provider
