+++ src/App.tsx
@@ -2,24 +2,23 @@ import { Box, ChakraProvider } from '@chakra-ui/react';
 import Content from './components/Content';
 import Navbar from './components/Navbar';
 import customizedTheme from './components/theme';
+import { CurrentContentProvider } from './contexts/CurrentContentContext';
 import './utilities/standard.css';
 
-/* App proper */
 const App: React.FC = () => {
   // TODO: Wrap in `SizeContext.Provider`?
   return (
     <ChakraProvider theme={customizedTheme}>
-      <Box
-        h="100%"
-        // bgColor={randomizedCSSrgb()}
-        bgColor="midnightblue"
-      >
-        {/* Navigation bar */}
-        <Navbar />
-
-        {/* Content section */}
-        <Content />
-      </Box>
+      <CurrentContentProvider>
+        <Box
+          h="100%"
+          // bgColor={randomizedCSSrgb()}
+          bgColor="midnightblue"
+        >
+          <Navbar />
+          <Content />
+        </Box>
+      </CurrentContentProvider>
     </ChakraProvider>
   );
 };
